First version of the opponent
We have the first version of the opponent. It first moves along the x
axis and then the y axis. We have no collision detection for it.
We also should implement path finding, so the snake can move
around its own tail in a intelligent way and perhaps find a more
effecient way to get that damn apple :) Now the fun begins! :D